Algumas funcionalidades implementadas

+++ src/main/java/com/casadocaminho/repositories/BeneficiarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.casadocaminho.models.Beneficiario;
 
-public interface BeneficiarioRepository extends CrudRepository<Beneficiario, Long> {
+public interface BeneficiarioRepository extends CrudRepository<Beneficiario, Integer> {
 
 }
