+++ src/main/java/com/casadocaminho/repositories/DoadorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.casadocaminho.models.Doador;
 
-public interface DoadorRepository extends CrudRepository<Doador, Long> {
+public interface DoadorRepository extends CrudRepository<Doador, Integer> {
 
 	public List<Doador> findByNome(String nome);
